BUG FIXED and logout features now work
Sorry for the previous push. There were a couple of bugs, but now they look to be fixed. I also added the sign out features to the buttons and the x button at the top right corner. All the Admin functions now work, but they are commented out right now in the "HomePageController" because they have another feature that I haven't added yet. Lots of bugs were fixed however, and that is the main point of this push.