FIXED VERSION of my last push
So My last version was super bug, so this is from the version before that. The search and book window fxml files are fixed. WARNING: I don't know how this is going to effect things since we already pushed a previous version of this code. Also the choice boxes work once again.